Transformed CommandRegisterer into interface, added impl for interface and some other things like getters in the main api class, so it does not need to be compiled
