+++ build.gradle
@@ -56,14 +56,9 @@ subprojects {
             name = "spigot-repo"
             url = "https://hub.spigotmc.org/nexus/content/groups/public/"
         }
-        maven {
-            name = "arim-mvn-lgpl3"
-            url = "https://mvn-repo.arim.space/lesser-gpl3/"
-        }
     }
 
     dependencies {
-        implementation("space.arim.morepaperlib:morepaperlib:0.4.3")
         implementation("net.kyori:adventure-api:4.3.0")
         implementation("net.kyori:adventure-platform-bukkit:4.3.0")
         implementation("net.kyori:adventure-text-serializer-legacy:4.13.1")
